pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-agent-tools"
version = "0.1.0"
description = "Local agent toolchain scaffold with PPT-first generation pipeline."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.99.0,<2.0",
  "pydantic>=2.9,<3.0",
  "python-pptx>=1.0.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
my-agent-tools = "my_agent_tools.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
